# Avoid allocating vector in message parsing