This fixes https://github.com/wokket/rust-hl7/issues/45.

I also noticed that this improves parsing performance marginally:

before: 
```
ORU parse               time:   [17.020 µs 17.097 µs 17.170 µs]
```
after:
```
ORU parse               time:   [16.822 µs 16.882 µs 16.948 µs]
```