@wokket 2023-09-20
Thanks for chipping in!

Are you using the library in anger?