@tvh 2023-09-21
No, I am not. I just stumbled over this looking for interesting rust projects to contribute to. We at @factisresearch have an internal implementation written in Haskell.